Refactor ProjectDiscovery into multiple files
